Feature/patrick product

+++ app/services/employee.py
@@ -106,11 +106,6 @@ class Employee:
                                bg="#4caf50", fg="white", cursor="hand2")
         search_button.place(x=420, y=10, width=150, height=31)
 
-        """style = ttk.Style()
-        style.configure("Search.TButton", background="green", foreground="white")
-        search_button = ttk.Button(frame, text="Search", style="Search.TButton", cursor="hand2")
-        search_button.place(x=420, y=10, width=150, height=31)"""
-
     def create_employee_details_frame(self):
         # Employee frame
         employee_frame = Frame(self.main_window, bd=3, relief=RIDGE, bg="white")
@@ -271,9 +266,10 @@ class Employee:
                     else:
                         confirm = messagebox.askyesno("Confirm", "Do you really want to delete?",
                                                       parent=self.main_window)
-                        cursor.execute("delete from Employee where EmployeeID=?", (self.employee_id.get(),))
-                        db_connection.commit()
-                        messagebox.showinfo("Delete", "Employee deleted successfully", parent=self.main_window)
+                        if confirm:
+                            cursor.execute("delete from Employee where EmployeeID=?", (self.employee_id.get(),))
+                            db_connection.commit()
+                            messagebox.showinfo("Delete", "Employee deleted successfully", parent=self.main_window)
                         self.clear_employee_data()
         except Exception as e:
             messagebox.showerror("Error", f"Error: {str(e)}", parent=self.main_window)
